+++ TooltipDecorator/util.js
@@ -95,6 +95,59 @@ const adjustDirection = function (tooltipDirection, overflow, rtl) {
 };
 
 /**
+ * Calculates the top and left position for `Tooltip` accounting for transform.
+ *
+ * @method
+ * @memberof limestone/TooltipDecorator
+ * @param   {Node} clientNode         	The value for client node
+ * @param   {String} tooltipDirection   Direction of tooltip
+ * @param   {Object} tooltipPosition    The `getBoundingClientRect` values for tooltip node
+ * @returns {Object}                    Tooltip's calculated new position accounting for transform
+ * @private
+ */
+const adjustTransform = function (clientNode, tooltipDirection, tooltipPosition) {
+	const componentNodes = clientNode?.childNodes.length ? [clientNode, ...clientNode.childNodes] : [clientNode];
+	const nodePositions = clientNode.getBoundingClientRect();
+	let hasTransform = false;
+	let transformMatrix = [];
+
+	componentNodes.forEach(node => {
+		if (window.getComputedStyle(node).transitionProperty === 'transform') {
+			hasTransform = true;
+			const matrix = window.getComputedStyle(node).transform.split('(')[1].split(')')[0].split(',');
+			transformMatrix = [
+				parseFloat(matrix[0]) - 1,
+				parseFloat(matrix[3]) - 1
+			];
+		}
+	});
+
+	if (!hasTransform) return tooltipPosition;
+
+	const horizontal = ((nodePositions.right - nodePositions.left) * transformMatrix[0]) / 2,
+		vertical = ((nodePositions.bottom - nodePositions.top) * transformMatrix[1]) / 2;
+
+	switch (tooltipDirection) {
+		case 'above':
+			tooltipPosition.top -= vertical;
+			break;
+		case 'below':
+			tooltipPosition.top += vertical;
+			break;
+		case 'left':
+			tooltipPosition.left -= horizontal;
+			break;
+		case 'right':
+			tooltipPosition.left += horizontal;
+			break;
+		default:
+			break;
+	}
+
+	return tooltipPosition;
+};
+
+/**
  * Calculates the overflow of `Tooltip` — if `Tooltip` is at the edge of the viewport.
  * Return the amount of overflow in a particular direction if there is overflow (false otherwise).
  *
@@ -233,8 +286,9 @@ const getLabelOffset = function (tooltipNode, tooltipDirection, tooltipPosition,
 };
 
 export {
-	adjustDirection,
 	adjustAnchor,
+	adjustDirection,
+	adjustTransform,
 	calcOverflow,
 	getLabelOffset,
 	getPosition