+++ VideoPlayer/VideoPlayer.js
@@ -2082,14 +2082,6 @@ const VideoPlayerBase = class extends Component {
 
 				{this.state.bottomControlsRendered ?
 					<div className={css.fullscreen} {...controlsAriaProps}>
-						<FeedbackContent
-							className={css.miniFeedback}
-							playbackRate={this.pulsedPlaybackRate || this.selectPlaybackRate(this.speedIndex)}
-							playbackState={this.pulsedPlaybackState || this.prevCommand}
-							visible={this.state.miniFeedbackVisible && !noMiniFeedback}
-						>
-							{secondsToTime(this.state.sliderTooltipTime, durFmt)}
-						</FeedbackContent>
 						{
 							this.state.mediaControlsVisible ?
 								<Button
@@ -2098,7 +2090,6 @@ const VideoPlayerBase = class extends Component {
 									icon="arrowhookleft"
 									iconFlip="auto"
 									onClick={this.handleBack}
-									size="small"
 								/> :
 								null
 						}
@@ -2111,6 +2102,14 @@ const VideoPlayerBase = class extends Component {
 								Only render when `this.state.mediaControlsVisible` is true in order for `Marquee`
 								to make calculations correctly in `MediaTitle`.
 							*/}
+							<FeedbackContent
+								className={css.miniFeedback}
+								playbackRate={this.pulsedPlaybackRate || this.selectPlaybackRate(this.speedIndex)}
+								playbackState={this.pulsedPlaybackState || this.prevCommand}
+								visible={this.state.miniFeedbackVisible && !noMiniFeedback}
+							>
+								{secondsToTime(this.state.sliderTooltipTime, durFmt, {includeHour: true})}
+							</FeedbackContent>
 							{this.state.mediaSliderVisible ?
 								<div className={css.infoFrame}>
 									<MediaTitle