+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/ActionGuide` styling to match the latest GUI
 - `limestone/IconItem` styling to match the latest GUI
 - `limestone/KeyGuide` styling to match the latest GUI
+- `limestone/VideoPlayer` styling to match the latest GUI
 
 ### Fixed