+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/IconItem` styling to match the latest GUI
 - `limestone/ImageItem` styling to match the latest GUI
 - `limestone/KeyGuide` styling to match the latest GUI
+- `limestone/Slider` styling to match the latest GUI
 - `limestone/TimePicker` styling to match the latest GUI
 - `limestone/VideoPlayer` styling to match the latest GUI